+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - fixed `destroy()` method + added test (#555)
   - added requirejs AMD support for `dist/parsley.js` and `dist/parsley.min.js`
     versions (#606)
+  - fixed custom namespace -multiple that didn't re-evaluated correctly (#595)
 
 ## 2.0.0-rc5